Split phonetic text into vovels & read separate text file
Splited into each vowels into a list, not sure if it is the right data type so would like to discuss further today or tomorrow maybe?

Also, commit changes for 17- read separate text file , now there is a new function to read every text file and return a string that could be passed to the "eng_to_ipa" function.